requirements.txt
streamlit==1.37.0
snowflake-snowpark-python==1.20.0
altair==5.3.0
